moving validation files to seperate folder to make code smaller/dryer. starting in messaging
